backend-service/requirements.txt
flask-login
flask-sqlalchemy
flask-cors
langchain==0.0.310
transformers==4.29.2
torch==2.1.0
sentencepiece==0.1.99
chromadb==0.4.13
openai==0.27.7
python-docx
sentence-transformers==2.2.2